Proof of concept for hardcoded node case of JUMP TO CANVAS ITEM
