+++ Stitch/Graph/Node/View/NodesView.swift
@@ -59,14 +59,7 @@ struct NodesView: View {
             //                        commentBoxes
             
             nodesOnlyView(nodePageData: nodePageData)
-//                .modifier(StitchUIScrollViewModifier())
         }
-//                       .modifier(StitchUIScrollViewModifier())
-        
-        
-//        nodesOnlyView(nodePageData: nodePageData)
-//            .modifier(StitchUIScrollViewModifier())
-        
            .modifier(CanvasEdgesViewModifier(document: document,
                                              graph: graph,
                                              graphUI: graphUI))
@@ -102,18 +95,6 @@ struct NodesView: View {
                       graph: graph,
                       graphUI: graphUI,
                       nodePageData: nodePageData)
-        .background {
-            GeometryReader { geometry in
-                Color.clear
-                    .onChange(of: geometry.frame(in: .named(WHOLE_GRAPH_COORDINATE_SPACE)),
-                              initial: true) { oldValue, newValue in
-                        log("SIZE READING: nodesOnlyView: WHOLE GRAPH frame: newValue.size: \(newValue.size)") // changes as we zoom?
-                        log("SIZE READING: nodesOnlyView: WHOLE GRAPH frame: newValue.mid: \(newValue.mid)")
-                        log("SIZE READING: nodesOnlyView: WHOLE GRAPH frame: newValue.origin: \(newValue.origin)")
-                        graph.graphUI.frameFromNodesOnlyView = newValue
-                    }
-            }
-        }
     }
 }
