+++ Stitch/Graph/Sidebar/Util/JumpToCanvasItem.swift
@@ -36,25 +36,70 @@ extension GraphState {
     @MainActor
     func panGraphToNodeLocation(id: CanvasItemId) {
         guard let canvasItem = self.getCanvasItem(id) else {
-            fatalErrorIfDebug("GraphState.sidebarItemTapped: no canvasItem found")
+            fatalErrorIfDebug("panGraphToNodeLocation: no canvasItem found")
             return
         }
+        
+        log("panGraphToNodeLocation: canvasItem \(id)")
 
         // location of canvasItem
         let position = canvasItem.position
         // ^^ should already exist when new group node created, because the child node already existed
+        
+        log("panGraphToNodeLocation: position \(position)")
 
         let newLocation = calculateMove(
             // size of GraphBaseView
             self.graphUI.frame,
             position)
+        
+        log("panGraphToNodeLocation: newLocation \(newLocation)")
+        
+        
 
         // TODO: how to slowly move over to the tapped layer? Using `withAnimation` on just graph offset does not animate the edges. (There's also a canvasItem text issue?)
         //        withAnimation(.easeInOut) {
-        self.graphMovement.localPosition = newLocation
-        self.graphMovement.localPreviousPosition = newLocation
+//        self.graphMovement.localPosition = newLocation
+//        self.graphMovement.localPreviousPosition = newLocation
         //        }
 
+//        self.graphUI.canvasJumpLocation = newLocation
+        
+        
+        let nodePosition = CGPoint(x: 1148, y: 645)
+//        
+//        let jumpPosition = CGPoint(
+////            x: 1148 + self.graphUI.frame.size.width/2,
+////            y: 645 + self.graphUI.frame.size.height/2
+//            
+//            
+//            // TODO: why do we have to SUBTRACT rather than add?
+//            x: 1148 - self.graphUI.frame.size.width/2,
+//            y: 645 - self.graphUI.frame.size.height/2
+//        )
+        // ^^ take view port into account
+        
+        // also ... scale ?
+        
+        let scale: CGFloat = self.documentDelegate?.graphMovement.zoomData.final ?? 1
+        
+        let jumpPosition = CGPoint(
+//            x: 1148 + self.graphUI.frame.size.width/2,
+//            y: 645 + self.graphUI.frame.size.height/2
+            
+            
+            // TODO: why do we have to SUBTRACT rather than add?
+            x: (1148 * scale) - self.graphUI.frame.size.width/2,
+            y: (645 * scale) - self.graphUI.frame.size.height/2
+        )
+        
+        log("panGraphToNodeLocation: scale: \(scale)")
+        log("panGraphToNodeLocation: jumpPosition: \(jumpPosition)")
+        
+        
+        
+        self.graphUI.canvasJumpLocation = jumpPosition
+        
         self.graphUI.selection = GraphUISelectionState()
         self.resetSelectedCanvasItems()
         canvasItem.select()
@@ -84,9 +129,9 @@ func calculateMove(_ graphViewFrame: CGRect,
     let newOffset = CGPoint(x: distance.x,
                             y: distance.y)
 
-    //    log("calculateMove: childPosition: \(childPosition)")
-    //    log("calculateMove: distance: \(distance)")
-    //    log("calculateMove: newOffset: \(newOffset)")
+    log("calculateMove: childPosition: \(childPosition)")
+    log("calculateMove: distance: \(distance)")
+    log("calculateMove: newOffset: \(newOffset)")
 
     return newOffset
 }
