+++ Stitch/Graph/Util/NodeSelection/NodeSelectionUtil.swift
@@ -168,37 +168,19 @@ extension GraphState {
             // Note: alternatively?: wipe this collection/set when gesure ends
             self.graphUI.nodesAlreadySelectedAtStartOfShiftNodeCursorBoxDrag = nil
         }
-        
-        // Necessary view frame data needed to determining selected nodes
-//        let zoom = 1 / self.graphMovement.zoomData.zoom
-//        let viewFrameSize = self.graphUI.frame.size
-//        let viewframeOrigin = CGPoint(x: -self.graphMovement.localPosition.x,
-//                                      y: -self.graphMovement.localPosition.y)
-//        let graphView = CGRect(origin: viewframeOrigin,
-//                               size: viewFrameSize)
-//        let scaledViewFrame = GraphState.getScaledViewFrame(scale: zoom,
-//                                                            graphView: graphView)
-        
-        // TODO: DEC 12: apparently no longer needed?
-//        guard let _selectionBoxInViewFrame: CGRect = GraphState
-//            .getScaledSelectionBox(selectionBox: selectionBox,
-//                                   scale: zoom,
-//                                   scaledViewFrameOrigin: scaledViewFrame.origin) else {
-//            return
-//        }
-        
+                
         let selectionBoxInViewFrame: CGRect = selectionBox
         
-        // log("processCanvasSelectionBoxChange: cachedSubviewData: selectionBoxInViewFrame.origin: \(selectionBoxInViewFrame.origin)")
-        // log("processCanvasSelectionBoxChange: cachedSubviewData: selectionBoxInViewFrame.size: \(selectionBoxInViewFrame.size)")
+         log("processCanvasSelectionBoxChange: cachedSubviewData: selectionBoxInViewFrame.origin: \(selectionBoxInViewFrame.origin)")
+         log("processCanvasSelectionBoxChange: cachedSubviewData: selectionBoxInViewFrame.size: \(selectionBoxInViewFrame.size)")
         
         for cachedSubviewData in self.visibleNodesViewModel.infiniteCanvasCache {
             let id = cachedSubviewData.key
             var cachedBounds = cachedSubviewData.value
             
-            // log("processCanvasSelectionBoxChange: cachedSubviewData: id: \(id)")
-            // log("processCanvasSelectionBoxChange: cachedSubviewData: cachedBounds.origin: \(cachedBounds.origin)")
-            // log("processCanvasSelectionBoxChange: cachedSubviewData: cachedBounds.size: \(cachedBounds.size)")
+             log("processCanvasSelectionBoxChange: cachedSubviewData: id: \(id)")
+             log("processCanvasSelectionBoxChange: cachedSubviewData: cachedBounds.origin: \(cachedBounds.origin)")
+             log("processCanvasSelectionBoxChange: cachedSubviewData: cachedBounds.size: \(cachedBounds.size)")
             
             guard self.visibleNodesViewModel.visibleCanvasIds.contains(id) else {
                 continue
